+++ designSystem/src/main/java/com/example/designsystem/AppTheme.kt
@@ -1,7 +1,5 @@
-package com.example.designsystem.textStyle
+package com.example.designsystem
 
-import MovioColors
-import MovioTextStyle
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
